checks: Improve tests
Closes: https://github.com/probabl-ai/skore/issues/3056
