+++ skore/tests/unit/checks/test_skd003_metrics_consistency.py
@@ -1,0 +1,36 @@
+import numpy as np
+from sklearn.datasets import make_classification
+from sklearn.linear_model import LinearRegression, LogisticRegression
+
+from skore import evaluate
+
+
+def test_passes_when_splits_are_consistent(regression_data):
+    """A CV-scoped check without findings appears as passed on the CV report."""
+    X, y = regression_data
+    report = evaluate(LinearRegression(), X, y, splitter=3)
+    passed_codes = set(report.checks.summarize().frame(section="passed")["code"])
+    assert "SKD003" in passed_codes
+
+
+def test_detects_inconsistent_splits():
+    """Check that the inconsistent performance across splits issue is detected."""
+    X, y = make_classification(n_samples=400, n_features=5, random_state=0)
+    report = evaluate(LogisticRegression(random_state=0), X, y, splitter=5)
+    assert "SKD003" not in set(report.checks.summarize().frame(section="issue")["code"])
+
+    # Corrupt the first split
+    y[0 : len(y) // 5] = np.random.RandomState(0).randint(0, 2, len(y) // 5)
+    report = evaluate(LogisticRegression(random_state=0), X, y, splitter=5)
+    issues = report.checks.summarize().frame(section="issue").set_index("code")
+    assert "SKD003" in issues.index
+    assert "split #0" in issues.loc["SKD003", "explanation"]
+    n_metrics = (
+        len(
+            report.metrics.summarize(data_source="test").frame(
+                aggregate=None, flat_index=True
+            )
+        )
+        - 2  # -2 for the timing metrics
+    )
+    assert f"for {n_metrics}/{n_metrics} metrics" in issues.loc["SKD003", "explanation"]
